# swagger.yaml
openapi: 3.0.0
info:
  title: Zendvo Auth API
  version: 1.0.0
  description: API for user registration and authentication

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      description: Creates a new user record with "unverified" status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "StrongP@ss123"
                name:
                  type: string
                  example: "John Doe"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      userId:
                        type: string
                        format: uuid
                      email:
                        type: string
        "400":
          description: Validation error (invalid email, weak password, missing fields)
        "409":
          description: Email already registered
        "413":
          description: Request body too large
        "429":
          description: Too many registration attempts
        "500":
          description: Internal server error
  /api/auth/forgot-password:
    post:
      summary: Request a password reset
      description: Sends a password reset email if the user exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        "200":
          description: Success response (always returned for security)
        "400":
          description: Invalid email format
        "429":
          description: Too many requests
        "500":
          description: Internal server error

  /api/auth/reset-password:
    post:
      summary: Reset user password
      description: Resets the password using a valid reset token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  format: uuid
                  example: "550e8400-e29b-41d4-a716-446655440000"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "NewStrongP@ss123"
      responses:
        "200":
          description: Password reset successfully
        "400":
          description: Invalid token or weak password
        "500":
          description: Internal server error
  /api/auth/login:
    post:
      summary: User login
      description: Returns access and refresh tokens for valid credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: "Password123!"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      user:
                        type: object
        "401":
          description: Invalid credentials
        "403":
          description: Account unverified or suspended
        "423":
          description: Account locked
        "429":
          description: Too many requests
        "500":
          description: Internal server error

  /api/auth/refresh:
    post:
      summary: Refresh access token
      description: Returns a new access token using a valid refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Tokens refreshed
        "401":
          description: Invalid or expired refresh token

  /api/auth/logout:
    post:
      summary: User logout
      description: Invalidates the provided refresh token.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Logged out successfully
        "400":
          description: Missing refresh token
        "401":
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT